Major refactoring!
