+++ TODO.md
@@ -5,10 +5,13 @@ See the [issues](https://github.com/drkostas/youbot/issues) too.
 - [X] Build YouTube Manager class
 - [X] Create child MySQL class
 - [X] Integrate YoutubeMysql class into the YoutubeManager class
+- [ ] Use the pypi packages I have created instead of the local ones
 - [ ] Create the workflow for the commenter
-- [ ] Roll the comments for each channel
+- [ ] Roll the comments for each channel - store comments in sql table?
+- [ ] Create table with errors
 - [ ] Create the workflow for the accumulator
 - [ ] Add SQL script for creating the tables needed
+- [ ] Send me email on fatal error
 - [ ] Recreate the Livestreaming module
 - [ ] Improve the YouTube api functions used (Activities api func - https://developers.google.com/youtube/v3/docs/activities/list)
 - [ ] Use multiple account (different api keys) to check for new comments
