+++ comments/sample_comments.txt
@@ -1,0 +1,4 @@
+Hey! I am a bot.
+Hello there!
+Nice video!
+Woo nice!
