+++ requirements.txt
@@ -1,16 +1,15 @@
 arrow~=1.1.1
-dropbox~=11.10.0
-gmail~=0.6.3
+bs4
 google-api-python-client~=2.7.0
 google-auth-oauthlib~=0.4.4
-jsonschema~=3.2.0
 httplib2~=0.19.1
-mysql-connector-python~=8.0.19
-mysql-connector~=2.2.9
 oauth2client~=4.1.3
 python-dateutil~=2.8.1
-PyYAML~=5.4.1
 requests~=2.25.1
-setuptools~=52.0.0
-termcolor~=1.1.0
-typer~=0.3.2
+setuptools>=52.0.0
+tqdm
+yaml-config-wrapper==1.0.4
+termcolor-logger==1.0.3
+high_sql==1.0.2
+pyemail_sender==1.0.1
+cloud_filemanager==1.0.1
