--- tests/test_data/test_configuration/actual_output_to_yaml.yml
@@ -1,7 +1,0 @@
-cloudstore:
-- subproperty1: 999
-  subproperty2:
-  - 123
-  - 234
-datastore: test
-tag: CHANGED VALUE
