--- tests/test_data/test_configuration/minimal_conf_correct.yml
@@ -1,7 +1,0 @@
-datastore: test
-cloudstore:
-  - subproperty1: 1
-    subproperty2:
-      - 123
-      - 234
-tag: test_tag
