--- tests/test_data/test_configuration/minimal_conf_wrong.yml
@@ -1,7 +1,0 @@
-datastore: test
-cloudstore:
-  - subproperty1: 10
-    subproperty2:
-      - 123
-      - 234
-tag: test_tag
