--- tests/test_data/test_configuration/template_conf.yml
@@ -1,13 +1,0 @@
-tag: production
-cloudstore:
-  - config:
-      api_key: apiqwerty
-    type: dropbox
-datastore:
-  - config:
-      hostname: host123
-      username: user1
-      password: pass2
-      db_name: db3
-      port: 3306
-    type: mysql
