--- youbot/cloudstore/__init__.py
@@ -1,7 +1,0 @@
-"""Cloudstore sub-package of YoutubeCommentBot."""
-
-from .dropbox_cloudstore import DropboxCloudstore
-
-__author__ = "drkostas"
-__email__ = "georgiou.kostas94@gmail.com"
-__version__ = "2.0"
