--- youbot/configuration/__init__.py
@@ -1,7 +1,0 @@
-"""Configuration sub-package of YoutubeCommentBot."""
-
-from .configuration import Configuration, validate_json_schema
-
-__author__ = "drkostas"
-__email__ = "georgiou.kostas94@gmail.com"
-__version__ = "2.0"
