--- youbot/configuration/yml_schema.json
@@ -1,17 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-04/schema#",
-  "title": "Python Configuration",
-  "description": "A json for python configuration in yml format",
-  "type": "object",
-  "properties": {
-    "tag": {
-      "type": "string"
-    }
-  },
-  "required": [
-    "tag"
-  ],
-  "definitions": {
-  },
-  "additionalProperties": true
-}
