--- youbot/datastore/__init__.py
@@ -1,7 +1,0 @@
-"""Cloudstore sub-package of YoutubeCommentBot."""
-
-from .mysql_datastore import YoutubeMySqlDatastore
-
-__author__ = "drkostas"
-__email__ = "georgiou.kostas94@gmail.com"
-__version__ = "2.0"
