--- youbot/emailer/__init__.py
@@ -1,7 +1,0 @@
-"""Emailer sub-package of YoutubeCommentBot."""
-
-from .gmail_emailer import GmailEmailer
-
-__author__ = "drkostas"
-__email__ = "georgiou.kostas94@gmail.com"
-__version__ = "2.0"
