--- youbot/fancy_logger/__init__.py
@@ -1,8 +1,0 @@
-"""FancyLog sub-package of YoutubeCommentBot."""
-
-from .colorized_logger import ColorizedLogger
-
-__author__ = "drkostas"
-__email__ = "georgiou.kostas94@gmail.com"
-__version__ = "2.0"
-
