--- youbot/youtube_utils/__init__.py
@@ -1,7 +1,0 @@
-"""Youtube Utils sub-package of YoutubeCommentBot."""
-
-from .youtube_manager import YoutubeApiV3, YoutubeManager
-
-__author__ = "drkostas"
-__email__ = "georgiou.kostas94@gmail.com"
-__version__ = "2.0"
